# api/openapi.yml
openapi: 3.0.0
paths:
...
  /holdings:
    get:
      operationId: HoldingsController_list
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HoldingResponseDto'
      security: &ref_1
        - vaultfolio_session: []
      summary: List the caller's holdings.
      tags: &ref_2
        - holdings
    post:
      operationId: HoldingsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CreateEtfHoldingRequestDto'
                - $ref: '#/components/schemas/CreateShareHoldingRequestDto'
                - $ref: '#/components/schemas/CreatePreciousMetalHoldingRequestDto'
                - $ref: '#/components/schemas/CreateCryptoHoldingRequestDto'
                - $ref: '#/components/schemas/CreateDepositMoneyHoldingRequestDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HoldingResponseDto'
        '400':
          description: One or more fields are invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HoldingValidationErrorResponseDto'
      security: *ref_1
      summary: Create a holding — request shape depends on `assetType`.
      tags: *ref_2
  /holdings/{id}:
    put:
      operationId: HoldingsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/UpdateEtfHoldingRequestDto'
                - $ref: '#/components/schemas/UpdateShareHoldingRequestDto'
                - $ref: '#/components/schemas/UpdatePreciousMetalHoldingRequestDto'
                - $ref: '#/components/schemas/UpdateCryptoHoldingRequestDto'
                - $ref: '#/components/schemas/UpdateDepositMoneyHoldingRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HoldingResponseDto'
        '400':
          description: One or more fields are invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HoldingValidationErrorResponseDto'
        '404':
          description: This holding no longer exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HoldingNotFoundErrorResponseDto'
      security: *ref_1
      summary: Update a holding — same shape as create, without `assetType`.
      tags: *ref_2
    delete:
      operationId: HoldingsController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Deleted.
        '404':
          description: This holding no longer exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HoldingNotFoundErrorResponseDto'
      security: *ref_1
      summary: Delete a holding.
      tags: *ref_2
  /accounts:
    get:
      operationId: AccountsController_listAll
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountSummaryDto'
        '403': &ref_3
          description: Caller is not an administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security: &ref_4
        - vaultfolio_session: []
      summary: List every account (admin only).
      tags: &ref_5
        - accounts
...
  /accounts/{id}:
    delete:
      operationId: AccountsController_deleteSelf
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Deleted.
        '403': *ref_3
        '409':
          description: Would leave no active administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsErrorResponseDto'
      security: *ref_4
      summary: Permanently delete an account past its retention window.
      tags: *ref_5
...
  /profile/account:
    delete:
      operationId: ProfileController_deleteAccount
      parameters: []
      responses:
        '204':
          description: Deleted.
        '409':
          description: Would leave no active administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileErrorResponseDto'
      security:
        - vaultfolio_session: []
      summary: Delete the caller's own account.
      tags: *ref_8
  /account-overview/accounts:
    get:
      operationId: AccountOverviewController_list
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountOverviewEntryDto'
      security: &ref_9
        - vaultfolio_session: []
      summary: List the caller's account-overview entries.
      tags: &ref_10
        - account-overview
    post:
      operationId: AccountOverviewController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountOverviewEntryRequestDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOverviewEntryDto'
        '400':
          description: One or more fields are invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOverviewValidationErrorResponseDto'
      security: *ref_9
      summary: Create an account-overview entry.
      tags: *ref_10
  /account-overview/accounts/{id}:
    put:
      operationId: AccountOverviewController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountOverviewEntryRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOverviewEntryDto'
        '400':
          description: One or more fields are invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOverviewValidationErrorResponseDto'
        '404':
          description: This account no longer exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOverviewNotFoundErrorResponseDto'
      security: *ref_9
      summary: >-
        Update an account-overview entry. Every field is optional; an omitted
        field is left unchanged.
      tags: *ref_10
    delete:
      operationId: AccountOverviewController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Deleted.
        '404':
          description: This account no longer exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOverviewNotFoundErrorResponseDto'
      security: *ref_9
      summary: Delete an account-overview entry.
      tags: *ref_10
info:
  title: Vaultfolio API
  description: >-
    Generated directly from the backend controllers/DTOs — see
    specs/031-openapi-swagger-integration. "Try it out" reuses the
    vaultfolio_session session cookie from a normal app login; it is not a
    separate credential scheme.
  version: 0.0.9
  contact: {}
...
servers: []
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: vaultfolio_session
  schemas:
...
    HoldingResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        assetType:
          type: string
          enum:
            - ETF
            - SHARE
            - PRECIOUS_METAL
            - CRYPTO
            - DEPOSIT_MONEY
        management:
          type: string
          description: Freeform label for who/what manages this holding.
        quantity:
          type: object
          nullable: true
          example: '10'
          description: Decimal string.
        purchasePrice:
          type: object
          nullable: true
          example: '123.45'
          description: Decimal string.
        purchaseDate:
          type: object
          nullable: true
          format: date
        isin:
          type: object
          nullable: true
        name:
          type: object
          nullable: true
        weightGrams:
          type: object
          nullable: true
          example: '15.5'
          description: Decimal string.
        currentValue:
          type: object
          nullable: true
          example: '1000.00'
          description: Decimal string.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - assetType
        - management
        - createdAt
        - updatedAt
    CreateEtfHoldingRequestDto:
      type: object
      properties:
        assetType:
          type: string
          enum: &ref_11
            - ETF
        management:
          type: string
        isin:
          type: string
        name:
          type: string
        quantity:
          type: string
          example: '10'
        purchasePrice:
          type: string
          example: '123.45'
      required:
        - assetType
        - management
        - isin
        - name
        - quantity
        - purchasePrice
    CreateShareHoldingRequestDto:
      type: object
      properties:
        assetType:
          type: string
          enum: &ref_12
            - SHARE
        management:
          type: string
        isin:
          type: string
        name:
          type: string
        quantity:
          type: string
          example: '10'
        purchasePrice:
          type: string
          example: '123.45'
        purchaseDate:
          type: string
          format: date
          description: Omit entirely, not "".
      required:
        - assetType
        - management
        - isin
        - name
        - quantity
        - purchasePrice
    CreatePreciousMetalHoldingRequestDto:
      type: object
      properties:
        assetType:
          type: string
          enum: &ref_13
            - PRECIOUS_METAL
        management:
          type: string
        name:
          type: string
        weightGrams:
          type: string
          example: '15.5'
        currentValue:
          type: string
          example: '1000.00'
          description: Used only by the distribution view.
      required:
        - assetType
        - management
        - name
        - weightGrams
    CreateCryptoHoldingRequestDto:
      type: object
      properties:
        assetType:
          type: string
          enum: &ref_14
            - CRYPTO
        management:
          type: string
        name:
          type: string
        quantity:
          type: string
          example: '0.5'
        purchasePrice:
          type: string
          example: '30000.00'
        purchaseDate:
          type: string
          format: date
          description: Omit entirely, not "".
      required:
        - assetType
        - management
        - name
        - quantity
        - purchasePrice
    CreateDepositMoneyHoldingRequestDto:
      type: object
      properties:
        assetType:
          type: string
          enum: &ref_15
            - DEPOSIT_MONEY
        management:
          type: string
        name:
          type: string
        currentValue:
          type: string
          example: '5000.00'
      required:
        - assetType
        - management
        - name
        - currentValue
    HoldingValidationErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          enum:
            - VALIDATION_FAILED
        message:
          type: string
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required:
        - error
        - message
        - fieldErrors
    UpdateEtfHoldingRequestDto:
      type: object
      properties:
        assetType:
          type: string
          enum: *ref_11
        management:
          type: string
        isin:
          type: string
        name:
          type: string
        quantity:
          type: string
          example: '10'
        purchasePrice:
          type: string
          example: '123.45'
      required:
        - assetType
        - management
        - isin
        - name
        - quantity
        - purchasePrice
    UpdateShareHoldingRequestDto:
      type: object
      properties:
        assetType:
          type: string
          enum: *ref_12
        management:
          type: string
        isin:
          type: string
        name:
          type: string
        quantity:
          type: string
          example: '10'
        purchasePrice:
          type: string
          example: '123.45'
        purchaseDate:
          type: string
          format: date
          description: Omit entirely, not "".
      required:
        - assetType
        - management
        - isin
        - name
        - quantity
        - purchasePrice
    UpdatePreciousMetalHoldingRequestDto:
      type: object
      properties:
        assetType:
          type: string
          enum: *ref_13
        management:
          type: string
        name:
          type: string
        weightGrams:
          type: string
          example: '15.5'
        currentValue:
          type: string
          example: '1000.00'
          description: Used only by the distribution view.
      required:
        - assetType
        - management
        - name
        - weightGrams
    UpdateCryptoHoldingRequestDto:
      type: object
      properties:
        assetType:
          type: string
          enum: *ref_14
        management:
          type: string
        name:
          type: string
        quantity:
          type: string
          example: '0.5'
        purchasePrice:
          type: string
          example: '30000.00'
        purchaseDate:
          type: string
          format: date
          description: Omit entirely, not "".
      required:
        - assetType
        - management
        - name
        - quantity
        - purchasePrice
    UpdateDepositMoneyHoldingRequestDto:
      type: object
      properties:
        assetType:
          type: string
          enum: *ref_15
        management:
          type: string
        name:
          type: string
        currentValue:
          type: string
          example: '5000.00'
      required:
        - assetType
        - management
        - name
        - currentValue
    HoldingNotFoundErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          enum:
            - HOLDING_NOT_FOUND
        message:
          type: string
      required:
        - error
        - message
    ErrorResponseDetailDto:
      type: object
      properties:
        field:
          type: string
          example: quantity
        message:
          type: string
          example: Must be a positive number.
      required:
        - field
        - message
    ErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          description: Stable machine-readable error code.
          example: not_found
        message:
          type: string
          example: Account not found.
        correlationId:
          type: string
          description: >-
            UUID for end-to-end tracing; matches the X-Correlation-Id response
            header.
          format: uuid
        details:
          description: >-
            Only present for field-level validation errors; non-empty when
            present.
          type: array
          items:
            $ref: '#/components/schemas/ErrorResponseDetailDto'
      required:
        - error
        - message
        - correlationId
    AccountSummaryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          example: member@example.com
        displayName:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - MEMBER
        status:
          type: string
          enum:
            - ACTIVE
            - ARCHIVED
        archivedAt:
          type: object
          format: date-time
          nullable: true
        retentionExpiresAt:
          type: object
          format: date-time
          nullable: true
        isLastActiveAdmin:
          type: boolean
        domainScopes:
          description: Domain ids this account is entitled to, independent of `role`.
          type: array
          items:
            type: string
      required:
        - id
        - email
        - displayName
        - role
        - status
        - isLastActiveAdmin
        - domainScopes
...
    AccountsErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          enum:
            - not_found
            - last_admin
            - already_archived
            - retention_expired
            - forbidden
            - invalid_domain
        message:
          type: string
      required:
        - error
        - message
...
    ProfileErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          enum:
            - invalid_display_name
            - email_unavailable
            - email_delivery_failed
            - invalid_token
            - invalid_password
            - invalid_current_password
            - last_admin
            - deletion_failed
            - invalid_email_language
            - bot_protection_failed
        message:
          type: string
      required:
        - error
        - message
...
    AccountOverviewEntryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
          enum: &ref_17
            - GENERAL
            - LEISURE
            - SAVINGS
            - DEPOT
            - CREDIT_CARD
            - OTHER
        status:
          type: string
          enum: &ref_18
            - ACTIVE
            - DECOMMISSIONED
          description: >-
            A decommissioned account stays listed, sorted after the active ones
            within its category.
        provider:
          type: object
          nullable: true
        website:
          type: object
          nullable: true
        purpose:
          type: object
          nullable: true
        cardUsage:
          type: object
          nullable: true
        requiredMinimum:
          type: object
          nullable: true
        notes:
          type: object
          nullable: true
        cardNumber:
          type: object
          nullable: true
          description: Full card number as entered (CREDIT_CARD accounts only).
        validUntil:
          type: object
          nullable: true
          example: 12/29
          description: MM/YY.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - status
        - createdAt
        - updatedAt
    CreateAccountOverviewEntryRequestDto:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: *ref_17
          description: Omit to default to 'OTHER'.
        status:
          type: string
          enum: *ref_18
          description: Omit to default to 'ACTIVE'.
        provider:
          type: string
        website:
          type: string
        purpose:
          type: string
        cardUsage:
          type: string
        requiredMinimum:
          type: string
        notes:
          type: string
        cardNumber:
          type: string
        validUntil:
          type: string
          example: 12/29
      required:
        - name
    AccountOverviewValidationErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          enum:
            - VALIDATION_FAILED
        message:
          type: string
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required:
        - error
        - message
        - fieldErrors
    UpdateAccountOverviewEntryRequestDto:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: *ref_17
          description: Omit to default to 'OTHER'.
        status:
          type: string
          enum: *ref_18
          description: Omit to default to 'ACTIVE'.
        provider:
          type: string
        website:
          type: string
        purpose:
          type: string
        cardUsage:
          type: string
        requiredMinimum:
          type: string
        notes:
          type: string
        cardNumber:
          type: string
        validUntil:
          type: string
          example: 12/29
      required:
        - name
    AccountOverviewNotFoundErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          enum:
            - ACCOUNT_NOT_FOUND
        message:
          type: string
      required:
        - error
        - message